# Replace SimpleSet with StringSet for performance