--- a/src/render.js
+++ b/src/render.js
@@ -72,11 +72,11 @@ module.exports = function render() {
     store._emitSelectionChange = false;
   }
 
-  if (store._deletedFeaturesToEmit.values().length) {
+  if (store._deletedFeaturesToEmit.length) {
     store.ctx.map.fire(Constants.events.DELETE, {
-      features: store._deletedFeaturesToEmit.values().map(feature => feature.toGeoJSON())
+      features: store._deletedFeaturesToEmit.map(feature => feature.toGeoJSON())
     });
-    store._deletedFeaturesToEmit.clear();
+    store._deletedFeaturesToEmit = [];
   }
 
   store.ctx.map.fire(Constants.events.RENDER, {});